requirements.txt
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
moviepy>=1.0.3
pydub>=0.25.1
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0